# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wealth Manager",
    "description": "Wealth Manager backend — source of truth for the API contract.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/assumptions": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Assumptions",
        "operationId": "get_assumptions_api_v1_assumptions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssumptionsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Put Assumptions",
        "operationId": "put_assumptions_api_v1_assumptions_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssumptionsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssumptionsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/inflation": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "List Inflation",
        "operationId": "list_inflation_api_v1_inflation_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InflationRateOut"
                  },
                  "type": "array",
                  "title": "Response List Inflation Api V1 Inflation Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Create Inflation",
        "operationId": "create_inflation_api_v1_inflation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InflationRateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InflationRateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/reports/pnl": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Pnl Report",
        "operationId": "pnl_report_api_v1_reports_pnl_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "group_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "class",
                    "account",
                    "asset"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group By"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PnlOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/net-worth": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Net Worth Report",
        "operationId": "net_worth_report_api_v1_reports_net_worth_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "both",
              "title": "Currency"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetWorthSeriesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/performance": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Performance Report",
        "operationId": "performance_report_api_v1_reports_performance_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/allocation": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Allocation Report",
        "operationId": "allocation_report_api_v1_reports_allocation_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "IRR",
              "title": "Currency"
            }
          },
          {
            "name": "goal_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/inflation-comparison": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Inflation Comparison",
        "operationId": "inflation_comparison_api_v1_reports_inflation_comparison_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InflationComparisonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/projection": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Projection Report",
        "operationId": "projection_report_api_v1_reports_projection_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "horizon_months",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 12,
              "title": "Horizon Months"
            }
          },
          {
            "name": "monthly_contribution",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                }
              ],
              "default": "0",
              "title": "Monthly Contribution"
            }
          },
          {
            "name": "scenario",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pessimistic",
                    "realistic",
                    "optimistic"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AllocationOut": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "total_value": {
            "type": "string",
            "title": "Total Value"
          },
          "current": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Current"
          },
          "target": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Target"
          },
          "drift": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Drift"
          },
          "rebalance": {
            "items": {
              "$ref": "#/components/schemas/RebalanceItem"
            },
            "type": "array",
            "title": "Rebalance"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "currency",
          "total_value",
          "current",
          "target",
          "drift",
          "rebalance"
        ],
        "title": "AllocationOut"
      },
...
      "AssumptionsOut": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "display_currency": {
            "$ref": "#/components/schemas/DisplayCurrency"
          },
          "hurdle_mode": {
            "$ref": "#/components/schemas/HurdleMode"
          },
          "hurdle_fixed_rate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hurdle Fixed Rate"
          },
          "growth_assumptions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Growth Assumptions"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "display_currency",
          "hurdle_mode"
        ],
        "title": "AssumptionsOut"
      },
      "AssumptionsUpdate": {
        "properties": {
          "display_currency": {
            "$ref": "#/components/schemas/DisplayCurrency",
            "default": "both"
          },
          "hurdle_mode": {
            "$ref": "#/components/schemas/HurdleMode",
            "default": "inflation"
          },
          "hurdle_fixed_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hurdle Fixed Rate"
          },
          "growth_assumptions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Growth Assumptions"
          }
        },
        "type": "object",
        "title": "AssumptionsUpdate"
      },
...
      "DisplayCurrency": {
        "type": "string",
        "enum": [
          "irr",
          "usd",
          "both"
        ],
        "title": "DisplayCurrency"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HurdleMode": {
        "type": "string",
        "enum": [
          "fixed",
          "inflation",
          "usd_growth"
        ],
        "title": "HurdleMode"
      },
...
      "InflationComparisonOut": {
        "properties": {
          "from": {
            "type": "string",
            "format": "date-time",
            "title": "From"
          },
          "to": {
            "type": "string",
            "format": "date-time",
            "title": "To"
          },
          "nominal_irr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nominal Irr"
          },
          "real_irr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Real Irr"
          },
          "usd_based": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usd Based"
          },
          "inflation": {
            "type": "string",
            "title": "Inflation"
          },
          "hurdle": {
            "additionalProperties": true,
            "type": "object",
            "title": "Hurdle"
          },
          "beats_inflation": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beats Inflation"
          },
          "beats_hurdle": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beats Hurdle"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to",
          "inflation",
          "hurdle"
        ],
        "title": "InflationComparisonOut"
      },
      "InflationRateCreate": {
        "properties": {
          "period_year": {
            "type": "integer",
            "maximum": 3000.0,
            "minimum": 1900.0,
            "title": "Period Year"
          },
          "period_month": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Period Month"
          },
          "rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Rate"
          },
          "source": {
            "type": "string",
            "maxLength": 64,
            "title": "Source",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "period_year",
          "period_month",
          "rate"
        ],
        "title": "InflationRateCreate"
      },
      "InflationRateOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "period_year": {
            "type": "integer",
            "title": "Period Year"
          },
          "period_month": {
            "type": "integer",
            "title": "Period Month"
          },
          "rate": {
            "type": "string",
            "title": "Rate"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "period_year",
          "period_month",
          "rate",
          "source",
          "created_at"
        ],
        "title": "InflationRateOut"
      },
...
      "NetWorthSeriesOut": {
        "properties": {
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/NetWorthPoint"
            },
            "type": "array",
            "title": "Series"
          }
        },
        "type": "object",
        "required": [
          "currency",
          "series"
        ],
        "title": "NetWorthSeriesOut"
      },
...
      "PerfMetrics": {
        "properties": {
          "xirr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xirr"
          },
          "twr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Twr"
          },
          "nominal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nominal"
          },
          "real": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Real"
          }
        },
        "type": "object",
        "title": "PerfMetrics"
      },
      "PerformanceOut": {
        "properties": {
          "from": {
            "type": "string",
            "format": "date-time",
            "title": "From"
          },
          "to": {
            "type": "string",
            "format": "date-time",
            "title": "To"
          },
          "irr": {
            "$ref": "#/components/schemas/PerfMetrics"
          },
          "usd": {
            "$ref": "#/components/schemas/PerfMetrics"
          },
          "inflation_cumulative": {
            "type": "string",
            "title": "Inflation Cumulative"
          },
          "usd_based": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usd Based"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to",
          "irr",
          "usd",
          "inflation_cumulative"
        ],
        "title": "PerformanceOut"
      },
      "PnlAmount": {
        "properties": {
          "irr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Irr"
          },
          "usd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usd"
          }
        },
        "type": "object",
        "title": "PnlAmount"
      },
      "PnlGroup": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "realized_irr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realized Irr"
          },
          "realized_usd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realized Usd"
          },
          "unrealized_irr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unrealized Irr"
          },
          "unrealized_usd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unrealized Usd"
          }
        },
        "type": "object",
        "required": [
          "key"
        ],
        "title": "PnlGroup"
      },
      "PnlOut": {
        "properties": {
          "realized": {
            "$ref": "#/components/schemas/PnlAmount"
          },
          "unrealized": {
            "$ref": "#/components/schemas/PnlAmount"
          },
          "group_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group By"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/PnlGroup"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "type": "object",
        "required": [
          "realized",
          "unrealized",
          "groups"
        ],
        "title": "PnlOut"
      },
...
      "ProjectionOut": {
        "properties": {
          "horizon_months": {
            "type": "integer",
            "title": "Horizon Months"
          },
          "monthly_contribution": {
            "type": "string",
            "title": "Monthly Contribution"
          },
          "scenarios": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ProjectionPoint"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Scenarios"
          }
        },
        "type": "object",
        "required": [
          "horizon_months",
          "monthly_contribution",
          "scenarios"
        ],
        "title": "ProjectionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}